Move middleware types
