+++ src/apply-middleware.ts
@@ -1,18 +1,5 @@
-import { InferActions, Producer } from "./types";
+import { Middleware, MiddlewareAction, Producer } from "./types";
 import { entries } from "./utils/entries";
-
-export type Middleware<T extends Producer<any, any> = Producer<unknown, Record<string, Callback>>> = (
-	producer: T,
-) => (done: NextMiddleware<InferActions<T>>) => NextMiddleware<InferActions<T>>;
-
-export type NextMiddleware<A> = (action: MiddlewareAction<A>) => unknown;
-
-export type MiddlewareAction<A = Record<string, unknown>> = {
-	[K in keyof A]: {
-		type: K;
-		arguments: Parameters<A[K]> extends [any, ...infer Rest] ? Rest : unknown[];
-	};
-}[keyof A];
 
 /**
  * Creates a producer enhancer that applies middleware to every dispatcher
@@ -20,8 +7,8 @@ export type MiddlewareAction<A = Record<string, unknown>> = {
  * @param middlewares The middleware to apply.
  * @returns The producer enhancer.
  */
-export function applyMiddleware<T extends Producer<any, any>>(...middlewares: Middleware<any>[]): (producer: T) => T;
-export function applyMiddleware(...middlewares: Middleware<any>[]) {
+export function applyMiddleware<T extends Producer<any, any>>(...middlewares: Middleware[]): (producer: T) => T;
+export function applyMiddleware(...middlewares: Middleware[]) {
 	return (producer: Producer<any, any>) => {
 		const dispatchers = producer.getDispatchers();
 		const chain = middlewares.map((middleware) => middleware(producer));
